// LDC.Atlas.WebApp/Client/app/shared/entities/cost-type.entity.ts
export interface CostType {
    costTypeCode: string;
    name: string;
    noAction: boolean;
    inPNL: boolean;
    nominalAccountCode: string;
    costTypeId: number;
    displayName: string;
    isACashCost: boolean;
    otherAcc: string;
    isATradeCost: boolean;
    isACommission: boolean;
    altCode: string;
    freight: boolean;
    insurance: boolean;
    noActualization: boolean;
    inPL: boolean;
    inStock: boolean;
    accrue: string;
    backOff: boolean;
    interfaceCode: string;
    sectionCode: string;
    objectCode: string;
}